Client: Handle grpc-status
Any non OK grpc status is stringified to not make a breaking API change. Ideally
this would not be the case the status can be useful for the application.

This commit also reverts the assumption that presence of grpc-message implies an
error. This is not according to the specification and even the server in this
repository sometimes sets grpc-message on OK. This didn't cause any issues
because the grpc-message was being lookup in the HTTP headers, while according
to the spec it must always be set in the trailers.